requirements.txt
robotpy-installer>=2019.0.0,< 2020.0.0
pyfrc>=2019.0.0,< 2020.0.0
robotpy-ctre>=2019.0.0,< 2020.0.0
robotpy-rev>=2019.0.0,< 2020.0.0